Split HTML terminal renderer tests
