+++ docs/DECISIONS.md
@@ -316,3 +316,4 @@
 - HTML chrome renderer coverage lives in `WorkspaceHTMLChromeRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Static HTML smoke coverage for primary regions, sidebar chrome, top-bar overflow, composer markup, context banners, runtime issues, and sidebar date buckets should stay together; tool-card, terminal, browser, secondary-pane, and review HTML coverage can be split into their own focused suites as those areas evolve.
 - HTML renderer architecture gates live in `ParityHTMLGateTests`, not the broad `ParityGateTests` catch-all. Pure HTML renderer delegation checks for tool cards, top bar, terminal, secondary panes, review, transcript, and sidebar should stay together; browser-specific HTML rendering stays in `ParityBrowserGateTests`, and mixed native/composer/workspace surface gates remain in the broad suite until they have enough focused domain coverage to split cleanly.
 - HTML tool-card renderer coverage lives in `WorkspaceHTMLToolCardRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Tool-card output, approval actions, artifact chips, image/document/appshot previews, and transcript ordering should stay beside the focused renderer so artifact UI work does not inflate the generic workspace surface suite.
+- HTML terminal renderer coverage lives in `WorkspaceHTMLTerminalRendererTests`, not the broad `WorkspaceSurfaceTests` catch-all. Visible terminal pane markup, command rows, clear controls, and running/stopped status labels should stay beside terminal HTML so terminal UI work can evolve without expanding generic surface projection tests.
